Implemented floating point arithmetic.

+++ src/main/scala/compilersandbox/parser/Node.scala
@@ -1,17 +1,17 @@
 package compilersandbox.parser
 
 sealed trait Node {
-  def compute(): Int
+  def compute(): Double
 }
 
 case class OperatorNode(operator: Operator, left: Node, right: Node) extends Node {
-  override def compute(): Int = {
+  override def compute(): Double = {
     operator.compute(left.compute(), right.compute())
   }
 }
 
 case class OperandNode(operand: Operand) extends Node {
-  override def compute(): Int = {
+  override def compute(): Double = {
     operand.value
   }
 }
