+++ src/main/scala/compilersandbox/parser/Operand.scala
@@ -1,3 +1,3 @@
 package compilersandbox.parser
 
-case class Operand(value: Int)
+case class Operand(value: Double)
